Added blob delete of exists

+++ etl/src/Piipan.Etl/Piipan.Etl.Func.BulkUpload/BulkUpload.cs
@@ -56,37 +56,43 @@ namespace Piipan.Etl.Func.BulkUpload
             log.LogInformation(myQueueItem);
             try
             {
-                var blockBlobClient =  _blobStream.Parse(myQueueItem, log);
-
-                Stream input = new System.IO.MemoryStream();
-
-                var response = blockBlobClient.DownloadTo(input);
-
-                var blobProperties = blockBlobClient.GetProperties().Value;
-
-                if (input != null)
-                {
-                    var participants = _participantParser.Parse(input);
-                    await _participantApi.AddParticipants(participants,  blobProperties.ETag.ToString())
-                            .ContinueWith(
-                                antecedent =>
-                                {
-                                    if (antecedent.Status == TaskStatus.RanToCompletion)
-                                    {
-                                        blockBlobClient.DeleteIfExists(DeleteSnapshotsOption.IncludeSnapshots);
-                                    }
-                                    else if (antecedent.Status == TaskStatus.Faulted)
-                                    {
-                                        log.LogError("Error inserting participants, blob not deleted.");
-                                        blockBlobClient.DeleteIfExists(DeleteSnapshotsOption.IncludeSnapshots);
-                                    }
-                                });
-                }
-                else
-                {
-                    // Can get here if Function does not have
-                    // permission to access blob URL
+                if(myQueueItem == null || myQueueItem.Length == 0){
+                    
                     log.LogError("No input stream was provided");
+                }
+                else {
+                    var blockBlobClient =  _blobStream.Parse(myQueueItem, log);
+
+                    Stream input = new System.IO.MemoryStream();
+
+                    var response = blockBlobClient.DownloadTo(input);
+
+                    var blobProperties = blockBlobClient.GetProperties().Value;
+
+                    if (input != null)
+                    {
+                        var participants = _participantParser.Parse(input);
+                        await _participantApi.AddParticipants(participants,  blobProperties.ETag.ToString())
+                                .ContinueWith(
+                                    antecedent =>
+                                    {
+                                        if (antecedent.Status == TaskStatus.RanToCompletion)
+                                        {
+                                            blockBlobClient.DeleteIfExists(DeleteSnapshotsOption.IncludeSnapshots);
+                                        }
+                                        else if (antecedent.Status == TaskStatus.Faulted)
+                                        {
+                                            log.LogError("Error inserting participants, blob not deleted.");
+                                            blockBlobClient.DeleteIfExists(DeleteSnapshotsOption.IncludeSnapshots);
+                                        }
+                                    });
+                    }
+                    else
+                    {
+                        // Can get here if Function does not have
+                        // permission to access blob URL
+                        log.LogError("No input stream was provided");
+                    }
                 }
             }
             catch (Exception ex)
